requirements.txt
chainlit
python-dotenv
tqdm
openai
langchain-google-genai